SnowMirror version 5.6.0
Performance API - SnowMirror Performance API provides a better performance compared to out-of-the-box REST or SOAP APIs. Performance API is a global-scoped application in ServiceNow. It achieves better download times by turning off ACL checks, turning off business rules and streaming the results. It is primarily targeted to the largest customers with the largest amounts of data. See Performance API documentation for details on how to get started.

The Performance API is now in its preview phase. It’s been thoroughly tested and is ready for use. We are eager to hear any feedback from our customers. The Performance API is available only in Cluster edition, but we are offering test licenses to anyone who is interested.