🔧 SnowMirror 5.10.3
Added the ability to define "custom columns to preserve" in the synchronization advanced settings page. These columns will not be deleted from the mirror table, even if they are not part of the synchronization.

Fixed: Synchronization filter did not show all matching tags when using "is not" condition.

Updated dependencies to their latest versions including Java 24.0.0+36.